[FEAT] Add email notifications for event registration
Implemented functionality to send an email notification upon user registration for an event. This includes creating a new job 'SendEventRegistrationEmail', a new mail 'EventRegistered', and a new view for the email content. Additionally, the job dispatch function has been added to 'AttachEventUserController' to trigger the email after the event is attached to the user.