[UPDATE] Refactor EventController and update Event Show Vue component
The EventController was refactored to improve handling of event participants and authentication checks. The associated Vue component, Event Show page, was updated to reflect these changes and improve user interactions. A clearer listing of event participants was added, and sign up links are now displayed based on user authentication state.