[FEAT] Revise layout and add registration link on Login page
Changes have been made to the layout of the Login.vue file to improve usability. A registration link has been added for users who do not yet have an account. Also, some adjustments are made to enhance the overall visual display of the page. The "Log in" button has been put in its own div for better structure.