[FEAT] Add reviews to event response in EventController
The update introduces the addition of event reviews to the event response in EventController. This addition ensures that whenever an event is sought, the associated reviews are also returned. This aids in giving a more comprehensive detail of each event by including user feedback which can be critical in decision making for other users.