[REFACTOR] Refactor EventController's store and edit methods
The store method in the EventController now also accepts title, description, and admin_id from the request input for an Event. The edit method now checks if the logged in user is authorized before they can edit an event. The Event model also includes admin_id in its fillable attribute.