[FEAT] Add popular categories to Welcome page
This update introduces a display of popular categories on the Welcome page. The 'WelcomeController' was adjusted to fetch the five most popular categories based on 'events_count'. These categories are passed through to the Welcome Vue.js component, which is updated to display each of the popular categories with a placeholder image and category name.